+++ src/Resources/app/administration/src/main.js
@@ -5,6 +5,8 @@ import './component/scalecommerce-vo-video-gallery';
 import './component/scalecommerce-vo-video-detail';
 import './module/sw-cms/elements/scalecommerce-vo-video';
 import './module/sw-cms/blocks/video/scalecommerce-vo-video';
+import './module/sw-cms/elements/scalecommerce-vo-media-split';
+import './module/sw-cms/blocks/video/scalecommerce-vo-media-split';
 
 Shopware.Service().register('scalecommerceVoApiService', (container) => {
     const initContainer = Shopware.Application.getContainer('init');