+++ src/Resources/app/administration/src/module/sw-cms/blocks/video/scalecommerce-vo-media-split/component/index.js
@@ -1,0 +1,5 @@
+import template from './scalecommerce-vo-media-split-block.html.twig';
+
+Shopware.Component.register('scalecommerce-vo-media-split-block', {
+    template,
+});