+++ src/Resources/app/administration/src/module/sw-cms/elements/scalecommerce-vo-media-split/preview/index.js
@@ -1,0 +1,5 @@
+import template from './scalecommerce-vo-media-split-preview.html.twig';
+
+Shopware.Component.register('scalecommerce-vo-media-split-preview', {
+    template,
+});